[S12P21E206-249] Fix: Redis와 DB 동기화 유지를 위한 쿠폰 + 유저 아이디 체크 로직 추가

+++ back/coupon/src/main/java/com/coupon/kafka/CouponKafkaConsumer.java
@@ -24,7 +24,7 @@ public class CouponKafkaConsumer {
     @KafkaListener(topics = "coupon", groupId = "coupon-v2")
     public void consume(CouponIssuedEvent event) {
         try {
-            Coupon coupon = couponRepository.findByCouponCode(event.couponCode().intValue())
+            Coupon coupon = couponRepository.findByCouponCode(event.couponCode())
                     .orElseThrow(() -> new CustomException(COUPON_NOT_FOUND));
 
             if (couponIssuedRepository.existsByUserIdAndCoupon(event.userId(), coupon)) {
